refactor(dashboard): extract misc domain (A1 phase 2 cont.)
- dashboard-misc.ts: excuse patterns / memory subsystems / keeper memory
  health / verification requests / TLA specs + TLC results / audit ledger.
  Split into two chunks during Edit (large block) — fetchVerificationRequests
  has opts.taskId (transcription-prone; verified against source).
- dashboard.ts: facade re-export only; consumers unchanged
- ripple import cleanup: post, AbortableRequestOptions (goals domain uses get only)
- typecheck green (tsc --noEmit)

dashboard.ts 1210 -> 837 lines after misc.
